cloud-functions/move-file-to-archive/requirements.txt
google-cloud-storage==1.19.1